refactor(indexer): use get_block_timestamp util fn
